requirements.txt
Django==4.0
djangorestframework==3.13.1
django-filter==21.1
psycopg2>=2.8
pandas==1.4.1